Improving code coverage. Extracting util functions from dom functions
