+++ uat-web-app/templates/alpha.html
@@ -37,6 +37,7 @@ $(document).ready(function () {
     localStorage.setItem("currentElement", "{{element}}");
 });
 </script>
+<script type="text/javascript" src="{{ url_for('static', filename='js/alphaUtils.js') }}"></script>
 <script type="text/javascript" src="{{ url_for('static', filename='js/alpha.js') }}"></script>
 <span id="topm"></span>
 <div class="container-fluid">
