+++ uat-web-app/templates/sorting.html
@@ -35,6 +35,7 @@
 
 <script type="text/javascript" src="{{ url_for('static', filename='js/sorting/captchaOperations.js') }}"></script>
 <script type="text/javascript" src="{{ url_for('static', filename='js/sorting/formOperations.js') }}"></script>
+<script type="text/javascript" src="{{ url_for('static', filename='js/sorting/mainUtils.js') }}"></script>
 <script type="text/javascript" src="{{ url_for('static', filename='js/sorting/main.js') }}"></script>
 
 <script type="text/javascript">
